fix: Resolve SwiftLint violations in ErrorReportingService
- Extract levelString(for:) helper to reduce cyclomatic complexity
- Extract sendToSentryLogger() helper for cleaner code organization
- Extract debugConsoleOutput() helper for DRY console logging
- Split domain-specific logging to ErrorReportingService+Logging.swift
  to fix file_length violation (was 530, now 419 lines)
- Make truncateErrorMessage() internal for extension access

Fixes: cyclomatic_complexity, function_body_length, file_length
